new makefile

+++ src/get_input.c
@@ -6,7 +6,7 @@
 /*   By: khansman <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/08/07 09:34:53 by khansman          #+#    #+#             */
-/*   Updated: 2016/09/10 16:14:08 by rlutsch          ###   ########.fr       */
+/*   Updated: 2016/09/11 11:30:57 by rlutsch          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,7 +62,7 @@ void	print_line(t_env *env)
 	l = -1;
 	k = ft_termsize(0);
 	CL_LINE;
-	l = (ft_strlen(I_L1) >= ft_termsize(0) - 5) ?
+	l = ((int)ft_strlen(I_L1) >= (int)ft_termsize(0) - 5) ?
 		ft_strlen(I_L1) + 5 - (ft_termsize(0) - 1) : -1;
 	(I_TMP2 == 0) ? ft_putstr(SH_L) : ft_putstr(SH_Q);
 	if (I_TMP2 == 0)
